+++ pkg/apis/validation/test/delete.go
